Revert "嗯"
This reverts commit 040f5dbb5d1ee680174ba1bac4a5c47dc6d0cd56.

+++ test1/Circle.cpp
@@ -1,6 +1,6 @@
 #include "Circle.h"
 
-Circle::Circle(int R)
+Circle::Circle(double R)
 {
 	cout << "Circle()" << endl;
 	m_dR = R;
